+++ src/index.ts
@@ -51,15 +51,21 @@ export = async function analyse(root = '.', opts: T.Options = {}) {
 
 			// Attempt to read gitignores
 			if (opts.checkIgnored) try {
-				let ignoresData = fs.readFileSync(folder + '.gitignore', { encoding: 'utf8' }); // may throw
+				let ignoresData = '';
+				try { ignoresData = fs.readFileSync(folder + '.gitignore', { encoding: 'utf8' }); }
+				catch { throw 404; }
 				const ignoresList = ignoresData.split(/\r?\n/).filter(line => line.trim() && !line.startsWith('#'));
 				const ignoredPaths = ignoresList.map(path => glob2regex('*' + path + '*', { extended: true }).source);
 				vendorData.push(...ignoredPaths);
-			} catch { }
+			} catch (err) {
+				if (err !== 404) throw err;
+			}
 
 			// Attempt to read gitattributes
 			if (opts.checkAttributes) try {
-				let attributesData = fs.readFileSync(folder + '.gitattributes', { encoding: 'utf8' }); // may throw
+				let attributesData = '';
+				try { attributesData = fs.readFileSync(folder + '.gitattributes', { encoding: 'utf8' }); }
+				catch { throw 404; }
 				// Custom vendor options
 				const vendorMatches = attributesData.matchAll(/^(\S+).*[^-]linguist-(vendored|generated|documentation)(?!=false)/gm);
 				for (const [_line, path] of vendorMatches) {
@@ -79,7 +85,9 @@ export = async function analyse(root = '.', opts: T.Options = {}) {
 					const fullPath = folder + convertToRegex(path).source.substr(1);
 					overrides[fullPath] = forcedLang;
 				}
-			} catch { }
+			} catch (err) {
+				if (err !== 404) throw err;
+			}
 
 		}
 	}